refactor: poster
