+++ src/features/map-poster/components/poster-board-filter.tsx
@@ -3,13 +3,13 @@
 import { Button } from "@/components/ui/button";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Label } from "@/components/ui/label";
-import type {
-  FilterStatus,
-  PosterBoardFilterState,
-} from "@/lib/hooks/usePosterBoardFilter";
 import type { Database } from "@/lib/types/supabase";
 import { ChevronDown, ChevronUp, Filter } from "lucide-react";
 import { useEffect, useState } from "react";
+import type {
+  FilterStatus,
+  PosterBoardFilterState,
+} from "../hooks/use-poster-board-filter";
 
 type BoardStatus = Database["public"]["Enums"]["poster_board_status"];
