+++ src/features/map-poster/config/status-config.ts
