+++ src/features/map-poster/hooks/use-poster-board-filter-optimized.ts
