+++ src/features/map-poster/services/poster-boards-stats.ts
