+++ src/features/map-poster/styles/poster-map-filter.css
