+++ src/features/map-poster/styles/poster-map.css
