+++ src/features/map-poster/types/poster-types.ts
