+++ src/features/map-poster/workers/poster-filter.worker.ts
